requirements.txt
pyTelegramBotAPI==4.33.0
requests==2.33.1
gspread==6.2.1
google-auth==2.49.2